+++ tests/timeseries/test_logical_meter.py
@@ -22,7 +22,7 @@ class TestLogicalMeter:
 
     async def test_grid_power_1(self, mocker: MockerFixture) -> None:
         """Test the grid power formula with a grid side meter."""
-        mockgrid = MockMicrogrid(grid_side_meter=True)
+        mockgrid = MockMicrogrid(grid_meter=True)
         mockgrid.add_batteries(2)
         mockgrid.add_solar_inverters(1)
         await mockgrid.start(mocker)
@@ -30,39 +30,40 @@ class TestLogicalMeter:
 
         grid_power_recv = logical_meter.grid_power.new_receiver()
 
-        main_meter_recv = get_resampled_stream(
+        grid_meter_recv = get_resampled_stream(
             logical_meter._namespace,  # pylint: disable=protected-access
-            mockgrid.main_meter_id,
+            mockgrid.meter_ids[0],
             ComponentMetricId.ACTIVE_POWER,
             Power.from_watts,
         )
 
         results = []
-        main_meter_data = []
+        grid_meter_data = []
         for count in range(10):
             await mockgrid.mock_resampler.send_meter_power(
                 [20.0 + count, 12.0, -13.0, -5.0]
             )
-            val = await main_meter_recv.receive()
+            val = await grid_meter_recv.receive()
             assert (
                 val is not None
                 and val.value is not None
                 and val.value.as_watts() != 0.0
             )
-            main_meter_data.append(val.value)
+            grid_meter_data.append(val.value)
 
             val = await grid_power_recv.receive()
             assert val is not None and val.value is not None
             results.append(val.value)
         await mockgrid.cleanup()
-        assert equal_float_lists(results, main_meter_data)
+        assert equal_float_lists(results, grid_meter_data)
 
     async def test_grid_power_2(
         self,
         mocker: MockerFixture,
     ) -> None:
         """Test the grid power formula without a grid side meter."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_consumer_meters(1)
         mockgrid.add_batteries(1, no_meter=False)
         mockgrid.add_batteries(1, no_meter=True)
         mockgrid.add_solar_inverters(1)
@@ -110,12 +111,13 @@ class TestLogicalMeter:
         assert len(results) == 10
         assert equal_float_lists(results, meter_sums)
 
-    async def test_grid_production_consumption_power(
+    async def test_grid_production_consumption_power_consumer_meter(
         self,
         mocker: MockerFixture,
     ) -> None:
         """Test the grid production and consumption power formulas."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_consumer_meters()
         mockgrid.add_batteries(2)
         mockgrid.add_solar_inverters(1)
         await mockgrid.start(mocker)
@@ -135,9 +137,34 @@ class TestLogicalMeter:
         assert (await grid_production_recv.receive()).value == Power.from_watts(4.0)
         assert (await grid_consumption_recv.receive()).value == Power.from_watts(0.0)
 
+    async def test_grid_production_consumption_power_no_grid_meter(
+        self,
+        mocker: MockerFixture,
+    ) -> None:
+        """Test the grid production and consumption power formulas."""
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_batteries(2)
+        mockgrid.add_solar_inverters(1)
+        await mockgrid.start(mocker)
+
+        logical_meter = microgrid.logical_meter()
+        grid_recv = logical_meter.grid_power.new_receiver()
+        grid_production_recv = logical_meter.grid_production_power.new_receiver()
+        grid_consumption_recv = logical_meter.grid_consumption_power.new_receiver()
+
+        await mockgrid.mock_resampler.send_meter_power([2.5, 3.5, 4.0])
+        assert (await grid_recv.receive()).value == Power.from_watts(10.0)
+        assert (await grid_production_recv.receive()).value == Power.from_watts(0.0)
+        assert (await grid_consumption_recv.receive()).value == Power.from_watts(10.0)
+
+        await mockgrid.mock_resampler.send_meter_power([3.0, -3.0, -4.0])
+        assert (await grid_recv.receive()).value == Power.from_watts(-4.0)
+        assert (await grid_production_recv.receive()).value == Power.from_watts(4.0)
+        assert (await grid_consumption_recv.receive()).value == Power.from_watts(0.0)
+
     async def test_chp_power(self, mocker: MockerFixture) -> None:
         """Test the chp power formula."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
         mockgrid.add_chps(1)
         mockgrid.add_batteries(2)
         await mockgrid.start(mocker)
@@ -151,7 +178,7 @@ class TestLogicalMeter:
             logical_meter.chp_consumption_power.new_receiver()
         )
 
-        await mockgrid.mock_resampler.send_meter_power([1.0, 2.0, 3.0, 4.0])
+        await mockgrid.mock_resampler.send_meter_power([2.0, 3.0, 4.0])
         assert (await chp_power_receiver.receive()).value == Power.from_watts(2.0)
         assert (
             await chp_production_power_receiver.receive()
@@ -160,7 +187,7 @@ class TestLogicalMeter:
             await chp_consumption_power_receiver.receive()
         ).value == Power.from_watts(2.0)
 
-        await mockgrid.mock_resampler.send_meter_power([-4.0, -12.0, None, 10.2])
+        await mockgrid.mock_resampler.send_meter_power([-12.0, None, 10.2])
         assert (await chp_power_receiver.receive()).value == Power.from_watts(-12.0)
         assert (
             await chp_production_power_receiver.receive()
@@ -171,7 +198,7 @@ class TestLogicalMeter:
 
     async def test_pv_power(self, mocker: MockerFixture) -> None:
         """Test the pv power formula."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
         mockgrid.add_solar_inverters(2)
         await mockgrid.start(mocker)
 
@@ -182,7 +209,7 @@ class TestLogicalMeter:
             logical_meter.pv_consumption_power.new_receiver()
         )
 
-        await mockgrid.mock_resampler.send_meter_power([10.0, -1.0, -2.0])
+        await mockgrid.mock_resampler.send_meter_power([-1.0, -2.0])
         assert (await pv_power_receiver.receive()).value == Power.from_watts(-3.0)
         assert (await pv_production_power_receiver.receive()).value == Power.from_watts(
             3.0
@@ -193,7 +220,7 @@ class TestLogicalMeter:
 
     async def test_pv_power_no_meter(self, mocker: MockerFixture) -> None:
         """Test the pv power formula."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
         mockgrid.add_solar_inverters(2, no_meter=True)
         await mockgrid.start(mocker)
 
@@ -215,7 +242,7 @@ class TestLogicalMeter:
 
     async def test_consumer_power_grid_meter(self, mocker: MockerFixture) -> None:
         """Test the consumer power formula with a grid meter."""
-        mockgrid = MockMicrogrid(grid_side_meter=True)
+        mockgrid = MockMicrogrid(grid_meter=True)
         mockgrid.add_batteries(2)
         mockgrid.add_solar_inverters(2)
         await mockgrid.start(mocker)
@@ -228,7 +255,8 @@ class TestLogicalMeter:
 
     async def test_consumer_power_no_grid_meter(self, mocker: MockerFixture) -> None:
         """Test the consumer power formula without a grid meter."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_consumer_meters()
         mockgrid.add_batteries(2)
         mockgrid.add_solar_inverters(2)
         await mockgrid.start(mocker)
@@ -239,9 +267,24 @@ class TestLogicalMeter:
         await mockgrid.mock_resampler.send_meter_power([20.0, 2.0, 3.0, 4.0, 5.0])
         assert (await consumer_power_receiver.receive()).value == Power.from_watts(20.0)
 
+    async def test_consumer_power_no_grid_meter_no_consumer_meter(
+        self, mocker: MockerFixture
+    ) -> None:
+        """Test the consumer power formula without a grid meter."""
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_batteries(2)
+        mockgrid.add_solar_inverters(2)
+        await mockgrid.start(mocker)
+
+        logical_meter = microgrid.logical_meter()
+        consumer_power_receiver = logical_meter.consumer_power.new_receiver()
+
+        await mockgrid.mock_resampler.send_non_existing_component_value()
+        assert (await consumer_power_receiver.receive()).value == Power.from_watts(0.0)
+
     async def test_producer_power(self, mocker: MockerFixture) -> None:
         """Test the producer power formula."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
         mockgrid.add_solar_inverters(2)
         mockgrid.add_chps(2)
         await mockgrid.start(mocker)
@@ -249,24 +292,43 @@ class TestLogicalMeter:
         logical_meter = microgrid.logical_meter()
         producer_power_receiver = logical_meter.producer_power.new_receiver()
 
-        await mockgrid.mock_resampler.send_meter_power([20.0, 2.0, 3.0, 4.0, 5.0])
+        await mockgrid.mock_resampler.send_meter_power([2.0, 3.0, 4.0, 5.0])
         assert (await producer_power_receiver.receive()).value == Power.from_watts(14.0)
 
     async def test_producer_power_no_chp(self, mocker: MockerFixture) -> None:
         """Test the producer power formula without a chp."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
         mockgrid.add_solar_inverters(2)
+
         await mockgrid.start(mocker)
 
         logical_meter = microgrid.logical_meter()
         producer_power_receiver = logical_meter.producer_power.new_receiver()
 
-        await mockgrid.mock_resampler.send_meter_power([20.0, 2.0, 3.0])
+        await mockgrid.mock_resampler.send_meter_power([2.0, 3.0])
         assert (await producer_power_receiver.receive()).value == Power.from_watts(5.0)
+
+    async def test_producer_power_no_pv_no_consumer_meter(
+        self, mocker: MockerFixture
+    ) -> None:
+        """Test the producer power formula without pv and without consumer meter."""
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_chps(1)
+        await mockgrid.start(mocker)
+
+        logical_meter = microgrid.logical_meter()
+        producer_power_receiver = logical_meter.producer_power.new_receiver()
+
+        # As there is only one meter in the microgrid, the formula interprets it
+        # as main meter instead of chp meter, so it reads the power from the
+        # chp component directly.
+        await mockgrid.mock_resampler.send_chp_power([2.0])
+        assert (await producer_power_receiver.receive()).value == Power.from_watts(2.0)
 
     async def test_producer_power_no_pv(self, mocker: MockerFixture) -> None:
         """Test the producer power formula without pv."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=False)
+        mockgrid.add_consumer_meters()
         mockgrid.add_chps(1)
         await mockgrid.start(mocker)
 
@@ -278,7 +340,7 @@ class TestLogicalMeter:
 
     async def test_no_producer_power(self, mocker: MockerFixture) -> None:
         """Test the producer power formula without producers."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
+        mockgrid = MockMicrogrid(grid_meter=True)
         await mockgrid.start(mocker)
 
         logical_meter = microgrid.logical_meter()